Add method and fix

+++ database/migrations/2021_09_14_084621_create_product_stock_table.php
@@ -18,7 +18,7 @@ class CreateProductStockTable extends Migration
             $table->text('ProductName');
             $table->text('Category');
             $table->integer('LotNo');
-            $table->integer('UnitPrice');
+            $table->float('UnitPrice');
             $table->integer('InStockAmt');
             $table->timestamp('RecordDate');
         });
